Refactor: 테스트 패키지 main과 같은 구조로 변경 (#49)
